+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Opening files after export now opens every exported group in a separate tab.
+- Grouped export filenames now remove wikilink markup from group names.
+
 ## [0.3.2] - 2026-09-06
 
 ### Added